# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: We Sell Houses API Documentation
  description: API for property management and communication between user and property agent
  contact:
    name: Filip Prosovsky
    email: prosovskyf@uni.coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://observe-prime.codio-box.uk/api/v1'
    description: Development server for this application api
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
...
    /login:
      summary: Login/Reset password
      description: >
        This resource manages login and reset password by sending token to email
      post:
        tags:
          - Agents
        description:
          Login
        security:
          - BasicAuth: []
        requestBody:
          description: Login user into application
          required: true
          content:
            application/json:
              schema:
                $ref: ./agents.json#/definitions/agentLogin
        responses:
          '200':
            description: User logged in, send back JWT and user role, id, username
            content:
              application/json:
                schema:
                  type: object
          '409':
            description: Account not activated, link to re-send activation link
            content:
              application/json:
                schema:
                  type: object
    /login/reset:
      post:
        tags:
          - Agents
        description:
          Reset password link
        requestBody:
          description: Username or email
          required: true
          content:
            application/json:
              schema:
                type: string
                properties:
                  user:
                    type: string
                    description: username/email
                required:
                  - user
        responses:
          '200':
            description: Link to reset password sent to email
          '404':
            description: Username/email does not exist
...
    /files:
      summary: Files
      description: >
        This resource manages files upload, get names and files deletion
        Checking if logged in user is agent with checkRole middleware
      get:
        parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
          description: URI path to file folder
        tags:
          - Files
        description:
          Get all file names in specific folder
        responses:
          '200':
            description: File names
            content:
              application/json:
                schema:
                  type: string
          '404':
            description: Not found
    /files/upload/category/image:
      put:
        tags:
          - Files
        description:
          Upload category image and store path to database
        parameters:
        - in: header
          name: category_id
          schema:
            type: integer
            description: Category ID
          required: true
        security:
          - BearerAuth: []
        requestBody:
          description: Picture
          required: true
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  image:
                    type: object
                    items:
                      type: string
                      format: binary
                required:
                  - image
        responses:
          '200':
            description: Picture uploaded
          '401':
            description: Not authorized
    /files/upload/property/images:
      put:
        tags:
          - Files
        description:
          Upload property images and store path to database
        parameters:
        - in: header
          name: property_id
          schema:
            type: integer
            description: Property ID
          required: true
        security:
          - BearerAuth: []
        requestBody:
          description: Pictures
          required: true
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  image:
                    type: object
                    items:
                      type: string
                      format: binary
                required:
                  - image
        responses:
          '200':
            description: Pictures uploaded
          '401':
            description: Not authorized
    /files/upload/property/video:
      put:
        tags:
          - Files
        description:
          Upload property video and store path to database
        parameters:
        - in: header
          name: property_id
          schema:
            type: integer
            description: Property ID
          required: true
        security:
          - BearerAuth: []
        requestBody:
          description: Video
          required: true
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  video:
                    type: object
                    items:
                      type: string
                      format: binary
                required:
                  - video
        responses:
          '200':
            description: Video uploaded
          '401':
            description: Not authorized
    /files/delete/category/image:
      delete:
        tags:
          - Files
        description:
          Delete category image and path from database
        security:
          - BearerAuth: []
        requestBody:
          description: Get category id, file name and path to the file
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    description: Category ID
                  name:
                    type: string
                    description: Category name
                  path:
                    type: string
                    description: path where file is to be deleted
                required:
                  - id
                  - name
                  - path
        responses:
          '200':
            description: Image deleted
          '401':
            description: Not authorized
    /files/delete/property/image:
      delete:
        tags:
          - Files
        description:
          Delete property images and path stored in database
        security:
          - BearerAuth: []
        requestBody:
          description: Get property id and file name
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    description: Property ID
                  imgName:
                    type: string
                    description: File name
                required:
                  - id
                  - imgName
        responses:
          '200':
            description: Image deleted
          '401':
            description: Not authorized
    /files/delete/property/video:
      delete:
        tags:
          - Files
        description:
          Delete property video and path stored in database
        security:
          - BearerAuth: []
        requestBody:
          description: Get property id and file name
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    description: Property ID
                  videoName:
                    type: string
                    description: File name
                required:
                  - id
                  - videoName
        responses:
          '200':
            description: Video deleted
          '401':
            description: Not authorized
...